Add workflow to regularly update the docker image (#17)
* Add workflow to regularly update VERSIONS file

The new workflow will check regularly whether a new release was tagged
on `dlang/dmd` and raise a PR to update the `VERSIONS.txt` if necessary.

Creating a PR instead of pushing to master ensures that a human still
verifies the updates generated by the script. Could be changed in the
future once the script has proven itself worthy.